requirements.txt
streamlit==0.87.0
requests==2.26.0
rembg==1.0.27
streamlit-lottie==0.1.1
Pillow==8.4.0
